Update patch set 3
Patch Set 3:

(Sorry for delays, I'll blame vacations, and this only being a side project.)

 > it needs to be serialised in the session, no?

It should certainly be available with other connection data. With serialization, what are you thinking of? SSL_SESSION_to_bytes_full or SerializeSecurityInfo?

The former is the only location which would make any sense in this review, but I cannot find any places where the deserialization happens, so not sure if that code is relevant?
The latter is necessarily required in any case, but I don't see how I can fit that into this review?

Patch-set: 3