Update patch set 3
Patch Set 3:

> It should certainly be available with other connection data. With
 > serialization, what are you thinking of? SSL_SESSION_to_bytes_full
 > or SerializeSecurityInfo?
 
SerializeSecurityInfo is an Chromium thing for passing SSL state across IPC boundaries. It doesn't have any interaction with the session cache.

If something's on the SSL_SESSION, it needs to get serialized in ssl_asn1.c. (The SESS_CERT attached isn't, but that's a bug. I just haven't gotten around to resolving that yet since SESS_CERT::cert_chain has some differences between client and server that I need to unravel.)

 > The former is the only location which would make any sense in this
 > review, but I cannot find any places where the deserialization
 > happens, so not sure if that code is relevant?
 > The latter is necessarily required in any case, but I don't see how
 > I can fit that into this review?

This isn't relevant for anything Chromium-based, but there's no reason why a client couldn't serialize and deserialize SSL_SESSIONs to implement its session cache. "This sometimes works and sometimes doesn't" is not a good API contract, so we should not rely on this.

Patch-set: 3