Update patch set 2
Patch Set 2: Code-Review+2

I think this is worthwhile because the |BN_CTX_get| trick is odd and, to an average reader, it makes the code look wrong. So I think it falls on the side of "too much magic".

Still, I'd like to keep the behaviour of |BN_CTX_get| in case we pull code from OpenSSL in the future that depends on it, or if outside code is depending on it.

Patch-set: 2
Label: Code-Review=+2