Update patch set 7
Patch Set 7:

> (4 comments)
 > 
 > Thanks for cleaning up the code :) Made a few more edits, based on
 > David's input.
 > 
 > One question for Adam though. You changed the default
 > (uninitialized) value of key_exchange_info from -1 to 0. The
 > variable might hold an enum, which might be 0 (at some time in the
 > future), and as there are anonymous cipers, a bit_strength value of
 > 0 might be confusing. Do you still prefer 0?

I think if we start needing more and more complex values in here, that would suggest that key_exchange_info is the wrong API and we'd want to ditch it in favor of something more structured per cipher suite type. I'd rather we not ad-hoc allow random values to be -1. (It's fine for now though. No need to overcomplicate things.)

I suppose using 0 does assume that no NamedCurve will ever be assigned with value 0. Dunno. I don't hugely care; I suppose if that ever changes, we can switch the internal representation of "not present" from 0 to -1 or whatever else later without changing the serialization. *shrug*

Patch-set: 7