Update patch set 2
Patch Set 2:

> get1_curves can certainly be dropped, but what is the proposed
 > answer to when servers want to accept only P-256 and curve25519,
 > but clients want to offer P-384 too?

I dunno, TBD mostly. :-) The early callback would work, but I suppose this API is actually just fine, so there's no sense in ditching it, even if we won't need it until later when Curve25519 exists. I'll change this to promote it to a proper function.

Patch-set: 2
Label: Code-Review=0