Adding more options for signing digest fallback.
Currently BoringSSL falls back to SHA-1 when we are unable to agree
on a signing digest between the server and client preferences. Since
there are cases where the client preferences are incomplete, we now
instead fallback to the server preferences.

This should not be merged in until the Chromium side is ready to be
committed.

BUG=468076

Change-Id: I996a2df213ae4d8b4062f0ab85b15262ca26f3c6