Stop using |ERR_peek_last_error| in crypto/ec.
Besides the fact that making decisions based on |ERR_peek_last_error|
is generally error prone, this was also interfering with the link-time
optimizer's ability to discard code in crypto/err during dead code
elimination.

Also, rename the output parameter of |BN_mod_sqrt| from |in| to |out| to
be consistent with the rest of the code.

Change-Id: I5ffa24c1c82ac18012b738236734946ffd2f0821