Update patch set 3
Patch Set 3:

Ok, I went with an output flag and _ex. I think it's slightly more fitting, although it's a not a clear thing. In any case, I'm guessing that you're more interested in picking one than in which is picked.

Please confirm that this works for you and I'll land it.

Patch-set: 3
Label: Code-Review=0