Adding more options for signing digest fallback.
Currently BoringSSL falls back to SHA-1 when we are unable to agree
on a signing digest between the server and client preferences. Since
there are cases where the client preferences are incomplete, we now
receive the complete list of digests that the client might support
in an ordered list based on preference. If the client does not
configure the list of digests supported by the private key, we fall
back to SHA1.

This should not be merged in until the Chromium side is ready to be
committed.

BUG=468076

Change-Id: I996a2df213ae4d8b4062f0ab85b15262ca26f3c6