Add an |SSL_CTX_ctrl| that calls |abort|.
The removal of this function from BoringSSL caused static Perforce
libraries in //third_party/perforce/p4sdk to fail to link against
BoringSSL.  Restoring this function allows linkage to occur.  TAP
global presubmit does not seem to encounter the call to |abort|.

(Of course, the use of OpenSSL's ABI is highly dubious, so this is not
in any sense a real fix.  The right solution is still for all users of
p4sdk to migrate away from it.  Or, somehow, to use real OpenSSL.)

Change-Id: Ic9c87107c84b492914cdae2e1c4d416699b1803c