Test that an ECDH point is on the curve.
This removes a sharp corner in the API where |ECDH_compute_key| assumed
that callers were either using ephemeral keys, or else had already
checked that the public key was on the curve.

A public key that's not on the curve can be in a small subgroup and thus
the result can leak information about the private key.

Change-Id: I77d10ce117b6efd87ebb4a631be3a9630f5e6636