BoringSSL: accept multi-prime keys with the wrong version.
(This is a commit of 103768685, which will likely not appear upstream.
It may be reverted when it is no longer needed.)

It seems that my OpenSSL code is generating multi-prime RSA keys with
the wrong version, but only BoringSSL is strict enough to notice.

This change makes BoringSSL accept these keys for now. A followup will
fix the OpenSSL code to generate them correctly too.

Change-Id: I93a5d2f4c84c2e8cc951d7cfd8ad5ae35f508527