Factor out checks that are common to all AEADs.
Factor out duplicate checks from AEAD implementations into the
higher-level functions that call them in order to make it easier to
create and audit AEAD implementations, and to better document how the
AEAD implementation works.

Some of the TLS/SSL-specific AEADs were not updated to use
|aead_seal_out_max_out_in_tag_len| and
|aead_open_out_max_out_in_tag_len| because those AEADs have overhead
other than the tag.

BUG=https://crbug.com/535758

Change-Id: I85d66fc8b1a20fb88211e7a3dc16a393e860a32e