Update patch set 2
Patch Set 2:

(1 comment)

One thing we possibly should care about is that, before this, these functions only failed on malloc failure. The failure mode of not checking would be that you might leak uninitialized data.

There is some code which doesn't check the return value, but almost all of it passes in a hard-coded value. There's just a little bit for which it's not immediately obviously not a concern: go/qulfr

If we're really worried, I suppose we could stick a memset in there. But I don't think we need to be. So this seems reasonable I think, apart from the comment on the test below.

Patch-set: 2
Label: Code-Review=0