Update patch set 4
Patch Set 4:

I addressed the test suggestion in patchset 4.

As far as the issue with consumers not checking the return value:

I will review the link you sent me. Another option if we are scared of this would be to continue treating iterations=0 the same as iterations=1 (do the computation and write the result), but return false on completion rather than true.

That said, consumers not checking the result code today seem like bugs to be fixed anyway.

Patch-set: 4
Label: Code-Review=0