Update patch set 4
Patch Set 4:

> I addressed the test suggestion in patchset 4.
 > 
 > As far as the issue with consumers not checking the return value:
 > 
 > I will review the link you sent me. Another option if we are scared
 > of this would be to continue treating iterations=0 the same as
 > iterations=1 (do the computation and write the result), but return
 > false on completion rather than true.

Oh yeah, that's also an option. But I'm not terribly scared of this. That code was just not obvious because of lots of layers. :-) If the iteration count is externally supplied without checking, you have problems most of the time anyway. I think this is basically only a concern for stuff like WebCrypto which is a really weird use case.

 > That said, consumers not checking the result code today seem like
 > bugs to be fixed anyway.

Probably. Although if you build with malloc failures fatal like a sane person, you can get away with not checking a lot of things.

Patch-set: 4