Update patch set 5
Patch Set 5:

In Patchset 5 I have switched to a more conservative approach:

The |out_key| is still filled like before (with a value corresponding to iterations==1), HOWEVER in addition false is returned.

This will avoid breaking any consumers that were subtly relying on that behavior, while allowing error-checking consumers to detect the failure.

I chose to do this after more carefully reviewing third party callers which do not check the value.

Another problem I found is callers that are passing "int" for the iteration count, in particular for JNI stuff (passing a jint). This is problematic for another reason in that it doesn't check for negatives and yet BoringSSL takes an unsigned value. I think historically OpenSSL took int for this parameter.

Patch-set: 5
Reviewer: Gerrit User 5445 <5445@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0