Ignore all extensions but renegotiation_info in SSL 3.0.
SSL 3.0 used to have a nice and simple rule around extensions. They don't
exist. And then RFC 5746 came along and made this all extremely confusing.

Rather than blocking ServerHello extension emission when
renegotiation_info is missing, block extension processing of all but
renegotiation_info extension in an SSL 3.0 server. Getting a mismatch
can break the handshake if the extension modifies the handshake.

See https://crbug.com/425979#c9 for a discussion of something similar
that came up in diagnosing https://poodle.io/'s buggy POODLE check.
This is analogous to upstream's
5a3d8eebb7667b32af0ccc3f12f314df6809d32d.

(Not supporting renego as a server in any form anyway, we may as well not
bother processing extensions, but then our extensions callbacks can't assume
the parse hooks are always called, so this seems somewhat better. All the
extensions will act as if the client didn't send them but still perform their
NULL contents handlers. And if we're still calling this function, we may as
well keep handling RI.)

Change-Id: Ie689a0e9ffb0369ef7a20ab4231005e87f32d5f8