Never use the internal session cache for a client.
The internal session cache is keyed on session ID, so this is completely
useless for clients (indeed we never look it up internally). Along the
way, tidy up ssl_update_mode to be more readable. The slight behavior
change is that SSL_CTX_add_session's return code no longer controls the
external callback as that makes little sense. (It can only fail on
allocation error. We only call it for new sessions, so the duplicate
case is impossible.)

The one thing of value it might have provided is managing the timeout;
the SSL_CTX_flush_sessions logic would flip the not_resumable bit and
cause us not to offer expired sessions (modulo SSL_CTX_flush_sessions's
delay and any discrepancies between the useless internal cache and the
external one). Instead, just check expiration when deciding whether or
not to offer a session.

This way clients that set SSL_SESS_CACHE_CLIENT blindly don't
accidentally consume gobs of memory.

BUG=531194

Change-Id: If97485beab21874f37737edc44df24e61ce23705