Don't bother sampling __func__.
Removing the function codes continued to sample __func__ for compatibility with
ERR_print_errors_cb, but not ERR_error_string_n. This probably isn't necessary
and we can just emit OPENSSL_internal for both.

This shaves 100kb from a release build of the bssl tool.

In doing so, put the unused function code back into ERR_put_error. We don't
need to pass an additional string in anymore, so OpenSSL compatibility with
anything which uses ERR_LIB_USER or ERR_get_next_error_library costs nothing.
(Not that we need it.)

Change-Id: If6af34628319ade4145190b6f30a0d820e00b20d