Remove curve check from tls12_check_peer_sigalg.
It only kicks on the server, where there isn't a curve list at all. This was
actually a mistake. https://boringssl-review.googlesource.com/1843 flipped it
from client-only to server-only.

But, either way, it also only kicks in for TLS 1.2, so one could bypass it by
just negotiating TLS 1.1. Upstream added it as part of Suite B mode, which
explains why it's 1.2 only. This is redundant with the certificate verifier, so
I think we can just ditch it. (We don't support anywhere near as many curves as
OpenSSL and things like enforcing minimum RSA key size are also left to the
certificate verifier.)

Change-Id: I7c12f2a32ba946a20e9ba6c70eff23bebcb60bb2