Update patch set 1
Patch Set 1:

(1 comment)

Attempted a butchered rewording of the commit comment.

I'm still a little unclear on the motivation for the original check, which I tried to capture in the description. Also, didn't TLS 1.3 make it fatal to ignore the SignatureAndHashAlgorithm in the SKE? I haven't caught up on the whole new key derivation scheme, so it may be entirely outdated for me to even mention that.

I'm fully inclined to let the caller (aka cert verifier) be responsible for all cryptographic policies and criteria. e.g. if it says "I want SHA-256" and the server sends SHA-1, the client should decide whether to yell or not. The concern I would have, though, is whether that means there's a security check that wouldn't happen - for example, if the caller sets a SigAndHashAlgs to disable SHA-1, the peer does a sig exchange with SHA-1, and the caller, not knowing the need to check, inadvertently picks up SHA-1. The existing code handles that (well, with the gross fallback to SHA-1), so it's unclear where the right line to draw for the API is.

That's the long-winded way of saying "I'm not sure; more info needed"

Patch-set: 1
Reviewer: Gerrit User 5190 <5190@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0